requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymupdf==1.23.8
openai==2.8.1
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.2